Fix and adds tests for evaluation of n=1 operators
EQ was broken; strncmp does not compare past first '\0' even when given
n is larger than the first such index.

AND/OR were broken; pointer arithmetic was performed before conversion
to pbg_expr_node*, which led to undefined behavior when dereferencing
each child to access its fields.

Added myriad tests for NOT, AND, OR, EQ, LT, GT, GTE, LTE. These are
not considered comprehensive.

GTE and LTE do not yet work. There is a parsing bug for operators of
length n=2. That will be fixed in a later commit.