Begin implementing 'flat' memory layout for AST
A standard, pointer-based tree implementation made for conceptually
easy code, but it did not, as it was, permit fast and easy resolution
of KEY literals during expression evaluation. This commit begins work
on a 'flat' memory layout in which node pointers are represented as
integers into two different arrays of nodes.

This implementation does not yet function. There are numerous memory
errors, and the current test suite leads to a segmentation fault. Code
is being committed in this state as I am moving to my Linux box to
access a wider suite of debugging tools, including valgrind.