Fix memory errors in pbg_parse
There were two primary issues with the previous commit.

First, pbg_free did not ensure _dynamicsz was offset by 1. This was
needed because _dynamicsz was initialized to 1 instead of 0 to ensure
indexing of KEY literals began at -1 instead of 0. (This itself is
needed as dynamic nodes are indexed with negative integers.) This bug
was fixed by intializing _dynamicsz to 0 and simply changing the code
in pbg_parse_r responsible for setting the index.

Second, the field length computation done in pbg_parse was not robust
to multiple consecutive closing parentheses. This was fixed by
introducing an 'open' flag to determine if a closing parentheses
actually closed an open, length-recording field.