Reimplement pbg_evalaute with 'flat' tree model
The old implementation of pbg_evaluate is re-added with minor tweaks
to get it working with the new tree representation. These primarily
involve changes to code accessing child nodes.

This commit also fixes a bug: KEY literals did not receive a copy of
their string representation. This is fixed by adding a call to strncmp,
the same way it is done with STRING literals.