Fix memory errors in previous commit
The initial reimplementation of pbg_evaluate has a few memory errors.
Most are fixed by using get_child instead of similar, inlined (but
erroneous) logic.

Prior to this commit, the EQ operation in pbg_evaluate produces a
memory error when comparing DATE literals. This is because C struct
alignment in combination with the use of an int and two char's in the
definition of pbg_date_type leads to two 'ghost' bytes not being
initialized. This is why understanding C struct byte alignment is
important! This problem is fixed in this commit by changing _MM and _DD
to be int instead of char. Memory optimization will be a priority later
on.

Removes unused include in pbg.h. This was initially added to make use
of the intptr_t type, but this is not needed in the 'flat'
representation.