Implement KEY resolution in pbg_evaluate
Changes prototype of pbg_evaluate to have dictionary return a
pbg_expr_node instead of a pbg_expr. This is necessary given the change
in roles of pbg_expr many commits before this one..

Adds code for resolving KEY literals and replacing dynamic array in the
pbg_expr instance provided to pbg_evaluate. To assist in cleanup, a new
local function is introduced: free_expr_node. In the previous commit,
it is clear that the same code for free'ing a pbg_expr_node was popping
up in multiple places. As this code is expected to change with the
introduction of new literals and operators, this logic was extracted
and traded for a function call.

Fixes implementation of the EXST operator's logic in pbg_evaluate_r. As
the dynamic array is replaced with each KEY lookup before the function
is invoked, there is no need to perform another dictionary lookup. A
single test for the type of the literal is all that is required. As
such, the dictionary no longer needs to be passed to pbg_evaluate_r, so
the prototype is changed accordingly.