Update code comments & add in-string parse rules
Adds a slew of new comments explaining the individual pieces of code in
function implementations. However, this commit does not include a full
explanation of the parsing algorithm used.

Reorganizes pieces of the code and adds headers and make code more
human-readable.

Adds ability for STRING literals to contain commas and escaped single
quotes.