Fix parsing bug that breaks nested operation eval
pbg_parse did not track every closing brace. It would only save the
last-most close. This is fixed by incrementing the index used to index
the closing array instead of just writing to the first index repeatedly.

Recognize newlines as whitespace characters. This allows an expression
string to be represented on multiple lines in a tree-like format. I'll
update print function to mirror this. After a game of Siege, of course.

Add stress tests to attempt to break the parser. It holds up quite well!

Fix error detection in test_evaluate by moving error instance reset code
into the check() macro. This allows syntatically broken tests to pass
when specified as such.

Add ERROR macro in pbg_test.h. -1 was being used previously, but this
makes the code a bit clearer.