Remove pbg_print
This library is meant to provide a minimal and efficient implementation
of PBG. pbg_print is outside of the scope of what this library is trying
to offer, and given the existence of pbg_gets, it has little utility.
Don't get me wrong: I don't have anything personal against it. It's just
business.