Move all helper functions from pbg.h to pbg.c
Any function that did not need to be in pbg.h is moved to pbg.c. This
was primarily done to declutter the name space and reduce the footprint
of the library. Functions can be re-added to pbg.h as necessary.

Modify the Makefile to add strict CFLAGS. Code now obeys these settings.
This library is intended to be simple and portable, so there is value in
ensuring the code is standard.

Pare down the codebase a bit. Clean up comments as well.