Remove comma separators from the grammar
Prior to this commit, commas were used to separate fields and
expressions. However, this came with a few downsides. First and
foremost, the strings in the grammar did not look very nice. Second,
commas required explicit support for whitespace insensitivity. Replacing
commas with any whitespace automatically comes with whitespace
insensitivity, which simplifies the parser.

Change tests.c to reflect the removal of commas from the grammar.

Add more syntax tests to test.c.