Add TYPE operator and type literals
The TYPE operator checks whether a set of arguments have a specified
type. For example, (@ DATE [a] [b] [c]) checks if the KEYs a, b, and c
all resolve to a DATE literal.

This operator is necessary as the resolution of a KEY may yield a
literal which causes an operator to throw an error. TYPE allows the user
to write the expression to guard against this, if it is a possibility
for the particular use case.

Add tests for the TYPE operator in test.c.

Add pbg_create_lt, and remove pbg_create_lt_true & pbg_create_lt_false.
The latter two functions had nearly identical implementations, and given
the introduction of type literals, it was easier to collapse the two
into a single function that takes the type as an argument.