Add example.c and rename "nodes" to "fields"
example.c is a small, complete example of how to compile and evaluate an
expression using the API.

After writing this example, it became apparent that users would benefit
from having access to functions for creating field literal types. This
is useful for implementing a dictionary, for example.

It also became apparent that a simpler method of checking if a function
returned an error would be to add a function to check this. As such,
pbg_iserror has been added to the API.

On the technical side of things, the introduction of the API field
literal factory functions led to the decomposition of the old
pbg_make_lt functions in pbg.c. The logic for inserting into the AST is
now encapsulated in pbg_store_static and pbg_store_dynamic, and the new
API functions are invoked and passed into either of these. It makes for
much cleaner code. Let's just hope gcc enjoys return value optimization.

Rename "nodes" to "fields". The former has been used to refer to the
nodes of the AST, and the latter has been used to refer to the
corresponding substring in the expression string. These two concepts are
no longer disjoint in name as both are now referred to as "fields"
throughout the API and implementation.