Rename and refactor code around variables
Static fields are now known as constant fields. Dynamic fields are now
known as variables. This change aligns the terminology of PBG with that
of boolean expressions, which will (hopefully) make understanding the
model a bit easier.

Rename KEY literal to a VAR literal. This was done to match the shift to
boolean expression terminology.

Rename PBG_UNKNOWN to PBG_NULL.

Refactor old comments to match new terminology. And just to keep things
updated.