Extend PBG to allow EQ,NEQ,LT,GT,LTE,GTE on BOOLs
The grammar has been extended to permit BOOL arguments to the
"arithmetic" operators. For example, (>(= 2 2)(!= 2 2)) can now be
evaluated to TRUE.

Add pbg_isbool to pbg.c to determine if a type is a BOOL. These types
include TRUE, FALSE, and any operator type. This permits the
implementation of the aforementioned grammar extension.

Add supporting tests for grammar change.

Update README.mc with changes to PBG.