Fix parser to ensure operator is first after open
The statement (TRUE) is not a valid PBG expression; however, the parser
recognized it as such. An explicit check for the type of the first
operator is added to pbg_parse to ensure it is an operator. If it is
not, a syntax error is generated and parsing halts.

Add pbg_unknown_type_err struct to represent data for an unknown type
error. This data is used to provide a printout of where the error
occurred in the input string, which in turn allows for more informative
errors dumps.

Add supporting tests.