Enforce inputs to operators are not NULL
pbg is designed to be readable, so errors should not be hidden via
evaluation to FALSE. For example, (> 10 NULL) should not be FALSE. FALSE
implies the GT operator received two values for which the first is not
greater than the second. However, 10 is not comparable to NULL, so
evaluating to FALSE would not comply with the contract defined by GT.

Similarly, (= 10 NULL) should not evaluate to FALSE, becausey that the values are not
equal. Evaluating this expression as FALSE would hide the fact that a
variable is not defined. Instead, the expression should explicitly check
if a variable is defined prior to usage if there is any doubt that it
may be defined.

Operator argument errors are thrown whenever an operator other than EXST
or TYPE receives NULL as an input.

Add supporting tests in test.c.