Reimplement pbg_parse
pbg_parse was originally written recursively with a helper function.
This was useful for testing, but it incurred unnecessary overhead. The
number of children was not computed beforehand, and the recursive calls
and numerous pointer dereferences weren't too dandy.

The code in this commit is NOT FUNCTIONAL. I am committing this as a
work in progress. I need to go to bed. There is a single bug remaining:
the number of children 'childi' is not re-pushed to the stack after each
update, so it is never updated. This leads to segmentation faults on
certain inputs. This will be fixed later.

Modify test.c to isolate single test that breaks this commit.