Fix new iterative pbg_parse implementation
The previous commit had a partially working iterative implementation of
pbg_parse. This commit fixes two bugs that previously caused some tests
to fail.

First, the stack operations in the third phase were fixed to push the
field ID and the corresponding input count at the same depth in the
stack. They were previously mismatched between depths in the stack. Bad
stuff. Details don't matter too much here. Summary: stack logic was a
mess previously. It has been amended to be much more readable and,
thankfully, functional. Two macros corresponding to stack peek
operations were added to clarify the code and make it easier to work
with.

Second, an explicit test was added for expressions without any fields.
This caught an error on the test expression "()" where errors occur if
numfields is 0.

Update test.c to run the entire test suite. It had been, in the prior
commit, reduced to a single representative test for the first bug fixed
here. This has been reverted.