Add pbg_parse_n to pbg.h
The old pbg_parse always required the length of the expression string to
be passed to it. This got annoying when a compile-time string was used
as strlen had to keep being called. Code got ugly.

Modify pbg_parse to no longer take a length argument.

Add pbg_parse_n. This one still takes a length argument.

Update example.c and test.c with the updated pbg_parse.